+++ people/admin.py
@@ -2,11 +2,11 @@
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 from cms.admin.placeholderadmin import PlaceholderAdminMixin
-from adminsortable.admin import SortableTabularInline
 
-from allink_core.allink_base.admin import AllinkBaseAdmin
+from allink_core.allink_base.admin import AllinkBaseAdmin, AllinkBaseAdminSortable
 from allink_core.allink_base.admin.forms import AllinkBaseAdminForm
-from allink_apps.people.models import PeopleImage, People
+from allink_apps.people.models import People
+
 
 @admin.register(People)
 class PeopleAdmin(PlaceholderAdminMixin, AllinkBaseAdmin):
@@ -23,13 +23,14 @@ class PeopleAdmin(PlaceholderAdminMixin, AllinkBaseAdmin):
                 'fields': (
                     'is_active',
                     ('firstname', 'lastname'),
-                    ('job_title', 'job_function'),
-                    ('email', 'website'),
-                    'company_name',
+                    # ('job_title', 'job_function'),
+                    ('email'),
+                    # ('email', 'website'),
+                    # 'company_name',
                     ('phone', 'mobile', 'fax'),
-                    ('street', 'street_nr'),
-                    'street_additional',
-                    ('place', 'zip_code'),
+                    # ('street', 'street_nr'),
+                    # 'street_additional',
+                    # ('place', 'zip_code'),
                     'slug',
                     'preview_image',
                 ),