+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The version numbers referring to the version in [allink_core](git@github.com/all
 ###### SETTINGS
 
 ###### TEMPLATES
+- people job_function (which it was used in teojakob for example) was substituted with property 'units'. You now have to add categories (with identifier='units') and tag th person with it. this allowes us to categories people without having to maintain both fields 'unit' and categories
 
 ###### URLS